Architecture
Spent a while figuring out how to structure / architect models and
services. Wanted to have all the logic in the model themselves, but had
a hard time injecting required services (endpointService,
storageService) into the model that is not a managed component of
angular. Decided instead to have separate Service manage logic and model
just have the data.